4.0.0/stability [4969] Implement submission comment requirements Publish dialog refactors progress (WIP): split components into their own files, typing & prop clean up. Pending full removal of setState prop from child theorethically-presentational components. [craftercms/craftercms#4969]
