+++ ui/app/src/models/GlobalState.ts
@@ -27,7 +27,6 @@ import { ErrorDialogStateProps } from '../components/SystemStatus/ErrorDialog';
 import { MinimizedDialogsStateProps } from './MinimizedDialog';
 import { NewContentDialogStateProps } from '../modules/Content/Authoring/NewContentDialog';
 import { HistoryDialogStateProps } from '../modules/Content/History/HistoryDialog';
-import { PublishDialogStateProps } from '../modules/Content/Publish/PublishDialog';
 import { DependenciesDialogStateProps } from '../modules/Content/Dependencies/DependenciesDialog';
 import { DeleteDialogStateProps } from '../modules/Content/Delete/DeleteDialog';
 import { EntityState } from './EntityState';
@@ -62,7 +61,7 @@ import { UnlockPublisherDialogStateProps } from '../components/UnlockPublisherDi
 import { WidgetDialogStateProps } from '../components/WidgetDialog';
 import { CodeEditorDialogStateProps } from '../components/CodeEditorDialog';
 import { SystemLinkId } from '../utils/system';
-import DateTimeFormatOptions = Intl.DateTimeFormatOptions;
+import { PublishDialogStateProps } from '../components/PublishDialog/utils';
 
 export interface PagedEntityState<T = any> extends EntityState<T> {
   page: any;
@@ -235,9 +234,10 @@ export interface GlobalState {
       error: ApiResponse;
       isFetching: boolean;
       localeCode: string;
-      dateTimeFormatOptions?: DateTimeFormatOptions;
+      dateTimeFormatOptions: Intl.DateTimeFormatOptions;
     };
     publishing: {
+      publishCommentRequired: boolean;
       deleteCommentRequired: boolean;
       bulkPublishRequired: boolean;
       publishByCommitRequired: boolean;
