+++ ui/app/src/state/actions/dialogs.ts
@@ -21,7 +21,6 @@ import { ViewVersionDialogStateProps } from '../../modules/Content/History/ViewV
 import { FetchContentVersion } from '../../models/Version';
 import { CompareVersionsDialogStateProps } from '../../modules/Content/History/CompareVersionsDialog';
 import { ConfirmDialogStateProps } from '../../components/Dialogs/ConfirmDialog';
-import { PublishDialogStateProps } from '../../modules/Content/Publish/PublishDialog';
 import { DeleteDialogStateProps } from '../../modules/Content/Delete/DeleteDialog';
 import { NewContentDialogStateProps } from '../../modules/Content/Authoring/NewContentDialog';
 import { DependenciesDialogStateProps } from '../../modules/Content/Dependencies/DependenciesDialog';
@@ -43,6 +42,7 @@ import { PublishingStatusDialogStateProps } from '../../components/PublishingSta
 import { UnlockPublisherDialogStateProps } from '../../components/UnlockPublisherDialog';
 import { WidgetDialogStateProps } from '../../components/WidgetDialog';
 import { CodeEditorDialogStateProps } from '../../components/CodeEditorDialog';
+import { PublishDialogStateProps } from '../../components/PublishDialog/utils';
 
 // region History
 export const showHistoryDialog = /*#__PURE__*/ createAction<Partial<HistoryDialogStateProps>>('SHOW_HISTORY_DIALOG');
